docker/requirements.txt
Flask==2.0.3 
werkzeug==2.0.3
mysql-connector-python==8.0.26
gunicorn==20.1.0